+++ R/utils-options.R
@@ -470,4 +470,5 @@ update_package_options <- function(debug = FALSE) {
   update_package_option("future.future.earlySignal", mode = "character", debug = debug)
   update_package_option("future.future.gc", mode = "character", debug = debug)
   update_package_option("future.future.local", mode = "character", debug = debug)
+  update_package_option("future.plan.earlySignal", mode = "character", debug = debug)
 }